RSA python implemetation
1- Generate two large prime numbers.
2- Use the generated prime numbers to find the Private and
the Public Keys of the RSA algorithm.
3- Use the Public keys to cipher a plain text file. (You should
open a plain text file, cipher the contents and save the
ciphered text in another file.)
4- Use the Private Keys to decipher a ciphered text file. (You
should open a ciphered text file, decipher the contents and
save the deciphered text in another file as a plain text.)
5- Your project should contain you own implementation for
the prime number generators, key generators,
cipher/decipher a text, GCD and the Modular
Exponentiation algorithms. Using build-in functions will
not be accepted.
6- Allowed programming languages are C/C++/Matlab.
7- The project is a solo project. Copying/sharing code is
completely not allowed.